refactor: Extract StoreButtons and use CSS for styles
